requirements.txt
Flask>=2.3.0
requests>=2.31.0
pywebview>=4.0.0
pystray>=0.19.0
pillow>=9.0.0
